Array based keyboard layout display.

+++ kbdlayout.h
@@ -11,7 +11,7 @@ class KBDLayoutCanvas: public wxPanel
 {
 private:
 	int scale = 40;
-	int act_col = -1, act_row = -1;
+	float act_col = -1, act_row = -1;
 	wxPanel *parent;
 
 public:
@@ -22,7 +22,7 @@ public:
 	void OnMouseDown(wxMouseEvent &event);
 	void OnMouseUp  (wxMouseEvent &event);
 
-	void DrawKey(wxDC &dc, wxString label, int x, int y, int w, int h);
+	void DrawKey(wxDC &dc, wxString label, float x, float y, float w, float h);
 	void SetScale(int s) { scale = s; }
 };
